+++ js/app/controllers/maincontroller.js
@@ -22,10 +22,14 @@ function ($rootScope, $scope, $timeout, $window, ArtistFactory,
 	// setup dark theme support for Nextcloud versions older than 25
 	OCA.Music.DarkThemeLegacySupport.applyOnElement(document.getElementById('app'));
 
-	// create a global rule to use themed icons for folders everywhere, the default icon-folder is not themed on NC 25 and later
-	const folderStyle = document.createElement('style');
-	folderStyle.innerHTML = `#app-view .icon-folder { background-image: url(${OC.MimeType.getIconUrl('dir')}) }`;
-	document.head.appendChild(folderStyle);
+	// Create a global rule to use themed icons for folders everywhere, the default icon-folder is not themed on NC 25 and later.
+	// It happens sometimes (at least on Chrome), that OC.MimeType is not yet present when we come here (see 
+	// https://github.com/owncloud/music/issues/1137). In those cases, we need to postpone registering the folder style.
+	OCA.Music.Utils.executeOnceRefAvailable(() => OC.MimeType, (ocMimeType) => {
+		const folderStyle = document.createElement('style');
+		folderStyle.innerHTML = `#app-view .icon-folder { background-image: url(${ocMimeType.getIconUrl('dir')}) }`;
+		document.head.appendChild(folderStyle);
+	});
 
 	$rootScope.playing = false;
 	$rootScope.playingView = null;